delete departure_test.go and move tests into fixtures_test.go
